requirements.txt
promptflow==1.16.2
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.12.0
azure-identity==1.19.0

